refact: move methods in appropriate files
