+++ pkg/controller/errors.go
@@ -1,0 +1,8 @@
+package controller
+
+type (
+	NoAnnotationError struct{ error }
+	AnnotationError   struct{ error }
+	RegistryError     struct{ error }
+	ClientError       struct{ error }
+)
