+++ pkg/controller/owned_secret_reconcilier.go
@@ -1,7 +1,12 @@
 package controller
 
 import (
+	"fmt"
+
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/klog"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
@@ -39,4 +44,32 @@ func (r *OwnedSecretReconcilier) Reconcile(req reconcile.Request) (reconcile.Res
 	default:
 		return reconcile.Result{RequeueAfter: requeueAfter}, err
 	}
+}
+
+// SynchronizeOwnedSecret duplicates the given secret in the given namespace.
+func SynchronizeOwnedSecret(ctx *Context, secret corev1.Secret, namespace string) error {
+	name := types.NamespacedName{Namespace: namespace, Name: secret.Name}
+	template := secret.DeepCopy()
+	template.ObjectMeta = metav1.ObjectMeta{
+		Name:        template.Name,
+		Namespace:   namespace,
+		Labels:      template.Labels,
+		Annotations: template.Annotations,
+		OwnerReferences: []metav1.OwnerReference{
+			{APIVersion: "v1", Kind: "Secret", Name: secret.Name, UID: secret.UID},
+		},
+	}
+	delete(template.Annotations, AllNamespacesAnnotation)
+	delete(template.Annotations, NamespaceSelectorAnnotation)
+
+	klog.V(3).Infof("update %T %s", secret, name)
+	err := ctx.client.Update(ctx, template)
+	if errors.IsNotFound(err) {
+		if err = ctx.client.Create(ctx, template); err != nil {
+			return ClientError{fmt.Errorf("failed to create %T %s: %w", secret, name, err)}
+		}
+	} else if err != nil {
+		return ClientError{fmt.Errorf("failed to update %T %s: %w", secret, name, err)}
+	}
+	return nil
 }
